+++ test/unit/fuzz-tools.test.js
@@ -223,6 +223,7 @@ import * as inflamV230 from '../../lib/inflam-v230.js';
 import * as prognosticV231 from '../../lib/prognostic-v231.js';
 import * as coagscoreV232 from '../../lib/coagscore-v232.js';
 import * as estimatorsV233 from '../../lib/estimators-v233.js';
+import * as dermscoreV234 from '../../lib/dermscore-v234.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -390,6 +391,7 @@ const MODULES = {
   'prognostic-v231.js': prognosticV231,
   'coagscore-v232.js': coagscoreV232,
   'estimators-v233.js': estimatorsV233,
+  'dermscore-v234.js': dermscoreV234,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];